fixed total_earn_from_points in update_loyalty API

+++ Client/models.py
@@ -414,7 +414,7 @@ class LoyaltyPoints(models.Model):
     amount_spend = models.FloatField(default=0, null=True, blank=True)
     number_points = models.PositiveIntegerField(default=0, null=True, blank=True)
     earn_points = models.FloatField(default=0, null=True, blank=True)
-    total_earn_from_points = models.PositiveIntegerField(default=0, null=True, blank=True)
+    total_earn_from_points = models.FloatField(default=0, null=True, blank=True)
     
     is_active = models.BooleanField(default=True)
     is_deleted = models.BooleanField(default=False)
